+++ Editor/translations/Editor_fr.ts
@@ -2536,6 +2536,16 @@ Do you want to run the Device Selector application to fix the problem?</source>
 Voulez-vous exécuter l’application Device Selector (Sélecteur de périphérique) pour résoudre le problème?</translation>
     </message>
     <message>
+        <location filename="../MainWindow.cpp" line="391"/>
+        <source>Device Selector</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../MainWindow.cpp" line="392"/>
+        <source>The Device Selector was not opened. It needs administrator approval to manage audio devices; you can try again anytime via File &gt; APO settings.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
         <location filename="../MainWindowParts/MainWindow.FileIO.cpp" line="61"/>
         <location filename="../MainWindowParts/MainWindow.FileIO.cpp" line="94"/>
         <location filename="../MainWindowParts/MainWindow.FileIO.cpp" line="100"/>